services/api-gateway/requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
sqlalchemy==2.0.34
pymysql==1.1.0

# --- Seguridad y autenticación ---
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.9.0

# --- Validación de datos ---
pydantic[email]==2.9.2
email-validator==2.2.0

# --- Variables de entorno ---
python-dotenv==1.0.1

# --- Comunicación entre servicios ---
httpx==0.27.2

# --- Utilidades opcionales (recomendado para debug) ---
python-multipart==0.0.9
